backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-environ==0.21.0
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dateutil==2.8.2
pytz==2023.3
